requirements.txt
twilio>=8.0.0
python-dotenv>=1.0.0
Flask>=2.0.0
requests>=2.0.0
gunicorn>=21.0.0
